Fix cast-qual in exported includes

+++ include/openssl/objects.h
@@ -117,32 +117,32 @@ const void *OBJ_bsearch_ex_(const void *key, const void *base, int num,
  * comparison routines do always not touch their arguments.
  */
 
-#define IMPLEMENT_OBJ_BSEARCH_CMP_FN(type1, type2, nm)                     \
-    static int nm##_cmp_BSEARCH_CMP_FN(const void *a_, const void *b_)     \
-    {                                                                      \
-        type1 const *a = a_;                                               \
-        type2 const *b = b_;                                               \
-        return nm##_cmp(a, b);                                             \
-    }                                                                      \
-    static type2 *OBJ_bsearch_##nm(type1 *key, type2 const *base, int num) \
-    {                                                                      \
-        return (type2 *)OBJ_bsearch_(key, base, num, sizeof(type2),        \
-            nm##_cmp_BSEARCH_CMP_FN);                                      \
-    }                                                                      \
+#define IMPLEMENT_OBJ_BSEARCH_CMP_FN(type1, type2, nm)                         \
+    static int nm##_cmp_BSEARCH_CMP_FN(const void *a_, const void *b_)         \
+    {                                                                          \
+        type1 const *a = a_;                                                   \
+        type2 const *b = b_;                                                   \
+        return nm##_cmp(a, b);                                                 \
+    }                                                                          \
+    static type2 *OBJ_bsearch_##nm(type1 *key, type2 const *base, int num)     \
+    {                                                                          \
+        return (type2 *)(uintptr_t)OBJ_bsearch_(key, base, num, sizeof(type2), \
+            nm##_cmp_BSEARCH_CMP_FN);                                          \
+    }                                                                          \
     extern void dummy_prototype(void)
 
-#define IMPLEMENT_OBJ_BSEARCH_GLOBAL_CMP_FN(type1, type2, nm)          \
-    static int nm##_cmp_BSEARCH_CMP_FN(const void *a_, const void *b_) \
-    {                                                                  \
-        type1 const *a = a_;                                           \
-        type2 const *b = b_;                                           \
-        return nm##_cmp(a, b);                                         \
-    }                                                                  \
-    type2 *OBJ_bsearch_##nm(type1 *key, type2 const *base, int num)    \
-    {                                                                  \
-        return (type2 *)OBJ_bsearch_(key, base, num, sizeof(type2),    \
-            nm##_cmp_BSEARCH_CMP_FN);                                  \
-    }                                                                  \
+#define IMPLEMENT_OBJ_BSEARCH_GLOBAL_CMP_FN(type1, type2, nm)                  \
+    static int nm##_cmp_BSEARCH_CMP_FN(const void *a_, const void *b_)         \
+    {                                                                          \
+        type1 const *a = a_;                                                   \
+        type2 const *b = b_;                                                   \
+        return nm##_cmp(a, b);                                                 \
+    }                                                                          \
+    type2 *OBJ_bsearch_##nm(type1 *key, type2 const *base, int num)            \
+    {                                                                          \
+        return (type2 *)(uintptr_t)OBJ_bsearch_(key, base, num, sizeof(type2), \
+            nm##_cmp_BSEARCH_CMP_FN);                                          \
+    }                                                                          \
     extern void dummy_prototype(void)
 
 #define OBJ_bsearch(type1, key, type2, base, num, cmp)                              \
